Resolved issue with images loading at different locations than they were supposed to be.  This could be reproduced by typing into the search field text and as it is loading one can scroll down or up by a fling where sometimes you notice images t\in the wrong place by gradually scrolling a page of images out of scope and then back into scope after all have been loaded and you find that images have been replaced by the ACTUAL image that was meant to be in that location.
In addition, synchronization was removed from ImageLoadAndDisplayInterface.java, since only the main thread is actually accessing the important ImageView binding and Bitmap setting methods for the operation.  This, along with a tag check resolving the forementioned issue above and a call to stop the previous animation, caused images to load much quicker.